+++ web/projects/marketplace/src/public-api.ts
@@ -5,11 +5,10 @@ export * from './components/links.component'
 export * from './components/dependencies.component'
 export * from './components/dependency.component'
 export * from './components/about.component'
-export * from './components/screenshots.component'
 export * from './components/flavors.component'
 export * from './components/versions.component'
 export * from './components/release-notes.component'
-export * from './components/aside.component'
+export * from './components/marketplace.component'
 export * from './components/store-icon.directive'
 export * from './components/registry.component'
 export * from './pipes/filter-packages.pipe'
