+++ web/projects/ui/src/app/routes/portal/routes/marketplace/marketplace.component.ts
@@ -2,27 +2,14 @@ import { CommonModule } from '@angular/common'
 import {
   ChangeDetectionStrategy,
   Component,
-  computed,
   inject,
   signal,
 } from '@angular/core'
 import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop'
 import { FormsModule } from '@angular/forms'
 import { ActivatedRoute, Router } from '@angular/router'
-import {
-  FilterPackagesPipe,
-  MarketplaceAsideComponent,
-} from '@start9labs/marketplace'
-import { i18nPipe, LocalizePipe } from '@start9labs/shared'
-import {
-  TuiCell,
-  TuiDataList,
-  TuiDropdown,
-  TuiScrollbar,
-  TuiTitle,
-} from '@taiga-ui/core'
-import { TuiAvatar, TuiChevron, TuiSelect, TuiSkeleton } from '@taiga-ui/kit'
-import { TuiCardLarge, TuiHeader } from '@taiga-ui/layout'
+import { MarketplaceComponent } from '@start9labs/marketplace'
+import { i18nPipe } from '@start9labs/shared'
 import { tap } from 'rxjs'
 import { ConfigService } from 'src/app/services/config.service'
 import { MarketplaceService } from 'src/app/services/marketplace.service'
@@ -35,62 +22,16 @@ import { MarketplaceTileComponent } from './components/tile.component'
 @Component({
   template: `
     <ng-container *title>{{ 'Marketplace' | i18n }}</ng-container>
-    <marketplace-aside
+    <marketplace
       [registry]="registry()"
       [(category)]="category"
       [(query)]="query"
     >
       <marketplace-registry-select />
-    </marketplace-aside>
-    <div class="content">
-      <header tuiHeader="h4">
-        <hgroup tuiTitle>
-          <h2 [tuiSkeleton]="!registry()">{{ categoryName() | localize }}</h2>
-        </hgroup>
-        <aside tuiAccessories>
-          <tui-textfield
-            tuiTextfieldSize="s"
-            tuiChevron
-            [tuiTextfieldCleaner]="false"
-          >
-            <input tuiSelect [(ngModel)]="sortLabel" />
-            <tui-data-list *tuiDropdown>
-              @for (key of sortKeys; track key) {
-                <button tuiOption [value]="getLabel(key)">
-                  {{ getLabel(key) }}
-                </button>
-              }
-            </tui-data-list>
-          </tui-textfield>
-        </aside>
-      </header>
-      <tui-scrollbar [style.flex]="1">
-        <section>
-          @if (registry(); as reg) {
-            @for (
-              pkg of reg.packages
-                | filterPackages: query() : category() : sort();
-              track $index
-            ) {
-              <button type="button" [marketplaceTile]="pkg"></button>
-            }
-          } @else {
-            @for (_ of '-'.repeat(15); track $index) {
-              <div tuiCardLarge="compact" [tuiSkeleton]="true">
-                <span tuiCell>
-                  <span tuiAvatar></span>
-                  <span tuiTitle>
-                    Loading
-                    <span tuiSubtitle>Loading</span>
-                  </span>
-                </span>
-                <span tuiDescription>Loading</span>
-              </div>
-            }
-          }
-        </section>
-      </tui-scrollbar>
-    </div>
+      <ng-template let-pkg>
+        <button type="button" [marketplaceTile]="pkg"></button>
+      </ng-template>
+    </marketplace>
   `,
   host: { class: 'g-page' },
   styles: `
@@ -100,73 +41,19 @@ import { MarketplaceTileComponent } from './components/tile.component'
       padding: 1px;
       background: #1c1d26;
     }
-
-    .content {
-      flex: 1;
-      min-width: 0;
-      display: flex;
-      flex-direction: column;
-      overflow: hidden;
-    }
-
-    [tuiHeader] {
-      align-items: center;
-      flex-wrap: wrap;
-      gap: 0.5rem 1rem;
-      padding: 1rem 2rem 0;
-    }
-
-    [tuiHeader] [tuiAccessories] {
-      flex-shrink: 0;
-      gap: 0;
-    }
-
-    tui-textfield {
-      inline-size: 10rem;
-    }
-
-    tui-scrollbar {
-      min-height: 0;
-    }
-
-    section {
-      padding: 1rem 2rem;
-      display: grid;
-      gap: 1rem;
-      grid-template-columns: repeat(auto-fill, minmax(17rem, 1fr));
-    }
-
-    :host-context(tui-root._mobile) [tuiHeader],
-    :host-context(tui-root._mobile) section {
-      padding-inline: 1rem;
-    }
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [
     CommonModule,
     FormsModule,
-    MarketplaceTileComponent,
-    TuiScrollbar,
-    FilterPackagesPipe,
     TitleDirective,
     i18nPipe,
-    LocalizePipe,
-    MarketplaceAsideComponent,
+    MarketplaceComponent,
     MarketplaceRegistrySelectComponent,
-    TuiChevron,
-    TuiDataList,
-    TuiDropdown,
-    TuiHeader,
-    TuiSelect,
-    TuiAvatar,
-    TuiSkeleton,
-    TuiCardLarge,
-    TuiCell,
-    TuiTitle,
+    MarketplaceTileComponent,
   ],
 })
-export default class MarketplaceComponent {
-  private readonly i18n = inject(i18nPipe)
+export default class Marketplace {
   private readonly marketplaceService = inject(MarketplaceService)
   private readonly configService = inject(ConfigService)
   private readonly router = inject(Router)
@@ -201,35 +88,9 @@ export default class MarketplaceComponent {
     )
     .subscribe()
 
-  readonly sort = signal('a')
-  readonly category = signal('all')
-  readonly query = signal('')
-  protected readonly sortKeys = ['a', '1']
+  protected readonly category = signal('all')
+  protected readonly query = signal('')
   protected readonly registry = toSignal(
     this.marketplaceService.currentRegistry$,
   )
-
-  protected get sortLabel(): string {
-    return this.getLabel(this.sort())
-  }
-
-  protected set sortLabel(label: string) {
-    this.sort.set(
-      this.sortKeys.find(key => this.getLabel(key) === label) || 'a',
-    )
-  }
-
-  protected readonly categoryName = computed(() => {
-    const category = this.category()
-    return this.registry()?.info?.categories?.[category]?.name || category
-  })
-
-  getLabel(sort: string) {
-    switch (sort) {
-      case 'a':
-        return this.i18n.transform('Alphabetical')
-      default:
-        return this.i18n.transform('Recently updated')
-    }
-  }
 }
